Refactor RegisterPage component to update UI and fix delete button in order menu
