+++ src/types/stallInfo.ts
@@ -1,0 +1,11 @@
+import { UUID } from "crypto";
+
+export type StallInfo = {
+  name: string;
+  commodities?: {
+    [key: UUID]: {
+      name: string;
+      price: number;
+    };
+  };
+};
